Add examples
