+++ examples/local_orb/07-pipek_mezey2.py
@@ -1,0 +1,80 @@
+#!/usr/bin/env python
+
+'''
+Pipek-Mezey (PM) localization and stability analysis
+
+This example demonstrates PM localization using both the CIAH and BFGS algorithms.
+It also illustrates stability analysis of the localized orbitals.
+'''
+
+import numpy
+from pyscf import gto, scf
+from pyscf.lib import logger
+from pyscf.lo import PipekMezey
+from pyscf.lo.tools import findiff_grad, findiff_hess
+
+mol = gto.Mole()
+mol.atom = '''
+O          0.00000        0.00000        0.11779
+H          0.00000        0.75545       -0.47116
+H          0.00000       -0.75545       -0.47116
+'''
+mol.basis = 'ccpvdz'
+mol.build()
+mf = scf.RHF(mol).run()
+
+log = logger.new_logger(mol, verbose=6)
+
+mo = mf.mo_coeff[:,:mol.nelectron//2]
+mlo = PipekMezey(mol, mo)
+
+# Validate gradient and Hessian against finite difference
+g, h_op, hdiag = mlo.gen_g_hop()
+
+h = numpy.zeros((mlo.pdim,mlo.pdim))
+x0 = mlo.zero_uniq_var()
+for i in range(mlo.pdim):
+    x0[i] = 1
+    h[:,i] = h_op(x0)
+    x0[i] = 0
+
+def func(x):
+    u = mlo.extract_rotation(x)
+    f = mlo.cost_function(u)
+    if mlo.maximize:
+        return -f
+    else:
+        return f
+
+def fgrad(x):
+    u = mlo.extract_rotation(x)
+    return mlo.get_grad(u)
+
+g_num = findiff_grad(func, x0)
+h_num = findiff_hess(fgrad, x0)
+hdiag_num = numpy.diag(h_num)
+
+log.info('Grad  error: %.3e', abs(g-g_num).max())
+log.info('Hess  error: %.3e', abs(h-h_num).max())
+log.info('Hdiag error: %.3e', abs(hdiag-hdiag_num).max())
+
+# localization + stability check using CIAH
+mlo.verbose = 4
+mlo.algorithm = 'ciah'
+mlo.kernel()
+
+while True:
+    mo, stable = mlo.stability(return_status=True)
+    if stable:
+        break
+    mlo.kernel(mo)
+
+# localization + Jacobi-based stability check using BFGS
+mlo.algorithm = 'bfgs'
+mlo.kernel()
+
+while True:
+    mo, stable = mlo.stability_jacobi(return_status=True)
+    if stable:
+        break
+    mlo.kernel(mo)
